push first combine

+++ main/src/main/java/me/nothing/login_/repository/StaffRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import me.nothing.login_.model.Staff;
 
 @Transactional
-public interface StaffRepository extends JpaRepository<Staff, Long> {
+public interface StaffRepository extends JpaRepository<Staff, Integer> {
 	@Query("SELECT u FROM Staff u WHERE u.username = :username")
 	public Staff findByUsername(@Param("username") String uername);
